+++ src/api/account.ts
@@ -244,7 +244,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorTypes: [ProcessorType.TOKEN_V2_PROCESSOR],
     });
     return getAccountOwnedTokens({
       aptosConfig: this.config,
@@ -276,7 +276,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorTypes: [ProcessorType.TOKEN_V2_PROCESSOR],
     });
     return getAccountOwnedTokensFromCollectionAddress({
       aptosConfig: this.config,
@@ -306,7 +306,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorTypes: [ProcessorType.TOKEN_V2_PROCESSOR],
     });
     return getAccountCollectionsWithOwnedTokens({
       aptosConfig: this.config,
@@ -357,7 +357,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorTypes: [ProcessorType.FUNGIBLE_ASSET_PROCESSOR],
     });
     return getAccountCoinsData({
       aptosConfig: this.config,
@@ -379,7 +379,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorTypes: [ProcessorType.FUNGIBLE_ASSET_PROCESSOR],
     });
     return getAccountCoinsCount({ aptosConfig: this.config, ...args });
   }
@@ -395,12 +395,7 @@ export class Account {
     accountAddress: AccountAddressInput;
     minimumLedgerVersion?: AnyNumber;
   }): Promise<number> {
-    await waitForIndexerOnVersion({
-      config: this.config,
-      minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
-    });
-    return getAccountCoinAmount({ aptosConfig: this.config, coinType: APTOS_COIN, ...args });
+    return this.getAccountCoinAmount({ coinType: APTOS_COIN, ...args });
   }
 
   /**
@@ -419,7 +414,7 @@ export class Account {
     await waitForIndexerOnVersion({
       config: this.config,
       minimumLedgerVersion: args.minimumLedgerVersion,
-      processorTypes: [ProcessorType.ACCOUNT_TRANSACTION_PROCESSOR],
+      processorTypes: [ProcessorType.FUNGIBLE_ASSET_PROCESSOR],
     });
     return getAccountCoinAmount({ aptosConfig: this.config, ...args });
   }